chore: relocated route builder helper function
