+++ app/client/src/sagas/ActionExecution/NavigateActionSaga.ts
@@ -5,7 +5,7 @@ import type { Page } from "@appsmith/constants/ReduxActionConstants";
 import AnalyticsUtil from "utils/AnalyticsUtil";
 import { getAppMode } from "@appsmith/selectors/applicationSelectors";
 import { APP_MODE } from "entities/App";
-import { getQueryStringfromObject } from "@appsmith/RouteBuilder";
+import { getQueryStringfromObject } from "@appsmith/entities/URLRedirect/URLAssembly";
 import history from "utils/history";
 import { setDataUrl } from "sagas/PageSagas";
 import AppsmithConsole from "utils/AppsmithConsole";
